+++ core/adapters/eth_tx.go
@@ -30,9 +30,6 @@ type EthTx struct {
 	DataFormat       string                  `json:"format"`
 	GasLimit         uint64                  `json:"gasLimit,omitempty"`
 
-	// GasPrice only needed for legacy tx manager
-	// GasPrice *utils.Big `json:"gasPrice" gorm:"type:numeric"`
-
 	// MinRequiredOutgoingConfirmations only works with bulletprooftxmanager
 	MinRequiredOutgoingConfirmations uint64 `json:"minRequiredOutgoingConfirmations,omitempty"`
 }
